requirements.txt
Django==5.2.4
django-allauth==65.10.0
django-ckeditor==6.7.3
python-dotenv==1.1.1
django-redis==6.0.0
bleach==6.2.0
tinycss2==1.4.0
django-crispy-forms==1.14.0
celery==5.5.3
django-celery-beat==2.8.1
django-filter==25.1

